refactor: Move Result into `<...>.util.rs` package
